Добавлен UserControl для отображения списка получателей
