+++ WpfMailSender/Data/DataBaseClass.cs
@@ -6,35 +6,35 @@ using System.Threading.Tasks;
 
 namespace WpfMailSender.Data
 {
-    //public class DataBaseClass
-    //{
-    //    private DataEmailsDataContext _emails = new DataEmailsDataContext();
-    //    public IQueryable<Emails> DBEmails
-    //    {
-    //        get
-    //        {
-    //            return _emails.Emails.Select(c => c);
-    //        }
-    //    }
+    public class DataBaseClass
+    {
+        //private DataEmailsDataContext _emails = new DataEmailsDataContext();
+        //public IQueryable<Emails> DBEmails
+        //{
+        //    get
+        //    {
+        //        return _emails.Emails.Select(c => c);
+        //    }
+        //}
 
-    //    private DataEmailsDataContext _smtpServers = new DataEmailsDataContext();
+        //private DataEmailsDataContext _smtpServers = new DataEmailsDataContext();
 
-    //    public IQueryable<Smtp> DBSmtp
-    //    {
-    //        get
-    //        {
-    //            return _smtpServers.Smtp.Select(c => c);
-    //        }
-    //    }
+        //public IQueryable<Smtp> DBSmtp
+        //{
+        //    get
+        //    {
+        //        return _smtpServers.Smtp.Select(c => c);
+        //    }
+        //}
 
-    //    private DataEmailsDataContext _senders = new DataEmailsDataContext();
+        //private DataEmailsDataContext _senders = new DataEmailsDataContext();
 
-    //    public IQueryable<Sender> DBSenders
-    //    {
-    //        get
-    //        {
-    //            return _senders.Sender.Select(c => c);
-    //        }
-    //    }
-    //}
+        //public IQueryable<Sender> DBSenders
+        //{
+        //    get
+        //    {
+        //        return _senders.Sender.Select(c => c);
+        //    }
+        //}
+    }
 }
