+++ WpfMailSender/ViewModels/WpfMailSenderViewModel.cs
@@ -14,18 +14,11 @@ namespace WpfMailSender.ViewModels
 {
     class WpfMailSenderViewModel : ViewModelBase
     {
+        public EmailInfoViewModel EmailInfoVM { get; }
         public MailSettings mailSettings { get; set; } = new MailSettings();
         EmailSendServiceClass _sendService;
         SchedulerClass _scheduler;
-        DataAccessService _dbAccessService;
-
-        private ObservableCollection<Emails> _emailsList;
-        public  ObservableCollection<Emails> EmailsList
-        {
-            get => _emailsList;
-            set => Set(ref _emailsList, value);
-        }
-
+        
 
         private Smtp _selectedSmtp;
         public Smtp SelectedSmtp
@@ -46,6 +39,7 @@ namespace WpfMailSender.ViewModels
 
         public WpfMailSenderViewModel()
         {
+            EmailInfoVM = new EmailInfoViewModel(this);
             #region Команды
 
             CloseApplicationCommand = new RelayCommand(OnCloseApplicationCommandExecuted, CanCloseApplicationCommandExecut);
@@ -53,20 +47,8 @@ namespace WpfMailSender.ViewModels
 
             #endregion
             
-            _dbAccessService = new DataAccessService();
-            GetEmails();
-
         }
 
-        void GetEmails()
-        {
-            EmailsList = new ObservableCollection<Emails>();
-            EmailsList.Clear();
-            foreach (var item in _dbAccessService.GetEmails())
-            {
-                EmailsList.Add(item);
-            }
-        }
 
         #region Команды
