+++ WpfMailSender/Views/EmailInfoView.xaml
@@ -1,0 +1,32 @@
+﻿<UserControl x:Class="WpfMailSender.Views.EmailInfoView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:WpfMailSender.Views"
+             xmlns:vm="clr-namespace:WpfMailSender.ViewModels"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800"
+             d:DataContext="{d:DesignInstance vm:EmailInfoViewModel}">
+    <DockPanel>
+        <StackPanel Orientation="Vertical" DockPanel.Dock="Top" Margin="5">
+            <Label FontWeight="Bold" Content="Список получателей" HorizontalAlignment="Center"/>
+            <StackPanel Orientation="Horizontal" Margin="0,5,0,0">
+                <Label Content="Поиск по имени" VerticalAlignment="Center"
+                       Margin="0,0,10,0"/>
+                <TextBox MinWidth="100" Height="30" Width="250"/>
+            </StackPanel>
+        </StackPanel>
+        <Button Content="Получить" DockPanel.Dock="Bottom"/>
+        <DataGrid AutoGenerateColumns="False" IsReadOnly="True"
+                  ItemsSource="{Binding EmailsList}"
+                  AlternatingRowBackground="AliceBlue">
+            <DataGrid.Columns>
+                <DataGridTextColumn Header="#" Width="10*" Binding="{Binding Id}"/>
+                <DataGridTextColumn Header="Имя" Width="50*" Binding="{Binding Name}"/>
+                <DataGridTextColumn Header="Email адрес" Width="100*" Binding="{Binding Email}"/>
+            </DataGrid.Columns>
+        </DataGrid>
+        
+    </DockPanel>
+</UserControl>
