+++ WpfMailSender/Views/WpfMailSender.xaml
@@ -7,6 +7,7 @@
         xmlns:xctk="http://schemas.xceed.com/wpf/xaml/toolkit" 
         xmlns:TabSwitcher1="clr-namespace:TabSwitcher;assembly=TabSwitcher"
         xmlns:vm="clr-namespace:WpfMailSender.ViewModels"
+        xmlns:view="clr-namespace:WpfMailSender.Views"
         mc:Ignorable="d"
         Title="Рассыльщик писем" Height="450" Width="800">
     <Window.DataContext>
@@ -124,17 +125,8 @@
                         </ToolBar>
                     </ToolBarTray>
 
-                    <DataGrid x:Name="dgEmails" 
-                              ItemsSource="{Binding EmailsList}"
-                              IsReadOnly="True" AutoGenerateColumns="False"
-                              AlternatingRowBackground="AliceBlue">
-                        <DataGrid.Columns>
-                            <DataGridTextColumn Header="#" Width="10*" Binding="{Binding Id}"/>
-                            <DataGridTextColumn Header="Имя" Width="50*" Binding="{Binding Name}"/>
-                            <DataGridTextColumn Header="Email адрес" Width="100*" Binding="{Binding Email}"/>
-                        </DataGrid.Columns>
-                    </DataGrid>
-                        
+                    <view:EmailInfoView DataContext="{Binding EmailInfoVM}"/>
+
                 </DockPanel>
             </TabItem>
