+++ WpfMailSender/packages.config
@@ -1,4 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="CommonServiceLocator" version="2.0.2" targetFramework="net472" />
   <package id="Extended.Wpf.Toolkit" version="3.8.1" targetFramework="net472" />
 </packages>
